{"schema":"decodex/commit/1","summary":"Split MCP autonomy proposal modules","authority":"XY-1115"}
